# Improve compile times by avoiding inlining