As discussed in #313, this change improves build times. I used the `simple` example, but made a copy of it such that it's `target` folder is independent of the one of this workspace. I then tested two kinds of changes: (a) just `touch main.rs` without changing anything and (b) change the name of the closure in `gl::load_with` to something else.

- `touch main.rs` rebuild: 
    - before: 3.9s
    - after: 2.45s
    - without `gl::load_with`: 2.4s (this line completely removed as a baseline; of course running the program would panic)
- change name of `symbol` in `load_with` closure (different number of chars):
    - before: 10.7s
    - after: 3.0s

So overall pretty nice wins I would say. Keep in mind though that these are no scientific benchmarks. I just repeated each "test" three times and calculated the approximate average in my head :grin: 

I guess one could optimize further by changing the public API. But this seems fine already and is not a breaking change. And by the way, as far as I can tell, the other generators do not have this problems or have this problem already fixed.

(And I hope it's OK to sneak the warning fix commit into this PR)